--- lib/util.js
@@ -1,2 +1,0 @@
-export const isDebugging = () =>
-  process.argv.includes('--debug') || process.env.NODE_ENV === 'DEBUG';
